perf: SIMD-accelerate SLoD/Dirichlet hot paths in katgpt-core
- slod: mobius_add_into scalar loop → SIMD fused_decay_write + fused_scale_acc
- slod: heat_kernel_weights_into inner accumulation → SIMD fused_scale_acc
- slod: Laplacian construction inner loop → SIMD copy+scale+scale_mul_inplace
- slod: frechet_mean weight_sum → SIMD sum instead of scalar iter().sum()
- slod: remove redundant sort_unstable_by after select_nth_unstable_by
- dirichlet: kv_cache_dirichlet_energy f64 round-trip → stay f32

+++ crates/katgpt-core/src/dirichlet.rs
@@ -99,6 +99,6 @@ pub fn kv_cache_dirichlet_energy(
     adjacency: &[(usize, usize)],
 ) -> (f32, f32) {
     let energy = dirichlet_energy(keys, kv_dim, adjacency);
-    let n_edges = adjacency.len().max(1) as f64;
-    (energy, (energy as f64 / n_edges) as f32)
+    let n_edges = adjacency.len().max(1) as f32;
+    (energy, energy / n_edges)
 }
